+++ tools/batools/pcommandmain.py
@@ -145,6 +145,8 @@ from batools.pcommands2 import (
     cst_test,
 )
 from batools.pcommands3 import (
+    test_game_run,
+    test_game_kill,
     compose_docker_gui_release,
     compose_docker_gui_debug,
     compose_docker_server_release,